ISSUE #1100: Store theme in database
