+++ src/Domain/Rewind/ActivityCountPerMonthChart.php
@@ -6,7 +6,7 @@ namespace App\Domain\Rewind;
 
 use App\Domain\Activity\SportType\SportType;
 use App\Domain\Calendar\Month;
-use App\Infrastructure\Theme\ChartColors;
+use App\Infrastructure\Theme\Theme;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
 use Symfony\Contracts\Translation\TranslatorInterface;
 
@@ -60,7 +60,7 @@ final readonly class ActivityCountPerMonthChart
                     'name' => $monthlyTotals[$month] ?? 0,
                     'value' => $monthlyActivityCounts[$key][$month] ?? 0,
                     'itemStyle' => [
-                        'color' => ChartColors::getColorForSportType($sportType),
+                        'color' => Theme::getColorForSportType($sportType),
                     ],
                 ];
             }
