+++ src/Domain/Rewind/DistancePerMonthChart.php
@@ -6,7 +6,7 @@ namespace App\Domain\Rewind;
 
 use App\Domain\Activity\SportType\SportType;
 use App\Domain\Calendar\Month;
-use App\Infrastructure\Theme\ChartColors;
+use App\Infrastructure\Theme\Theme;
 use App\Infrastructure\ValueObject\Measurement\Length\Kilometer;
 use App\Infrastructure\ValueObject\Measurement\UnitSystem;
 use App\Infrastructure\ValueObject\Time\SerializableDateTime;
@@ -69,7 +69,7 @@ final readonly class DistancePerMonthChart
                     'name' => $monthlyTotals[$month] ?? 0,
                     'value' => $monthlyDistances[$key][$month] ?? 0,
                     'itemStyle' => [
-                        'color' => ChartColors::getColorForSportType($sportType),
+                        'color' => Theme::getColorForSportType($sportType),
                     ],
                 ];
             }
