+++ src/Infrastructure/KeyValue/Key.php
@@ -8,4 +8,5 @@ enum Key: string
 {
     case STRAVA_GEAR_IMPORT = 'stravaGearImport';
     case ATHLETE = 'athlete';
+    case THEME = 'theme';
 }
