+++ src/Kernel.php
@@ -3,6 +3,8 @@
 namespace App;
 
 use App\Infrastructure\DependencyInjection\AppExpressionLanguageProvider;
+use App\Infrastructure\KeyValue\KeyValueStore;
+use App\Infrastructure\Theme\Theme;
 use Symfony\Bundle\FrameworkBundle\Kernel\MicroKernelTrait;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\Kernel as BaseKernel;
@@ -15,5 +17,13 @@ class Kernel extends BaseKernel
     {
         parent::build($container);
         $container->addExpressionLanguageProvider(new AppExpressionLanguageProvider());
+    }
+
+    protected function initializeContainer(): void
+    {
+        parent::initializeContainer();
+        /** @var KeyValueStore $keyValueStore */
+        $keyValueStore = $this->getContainer()->get(KeyValueStore::class);
+        Theme::setKeyValueStore($keyValueStore);
     }
 }
